requirements.txt
streamlit>=1.44.0
langchain>=0.3.24
langchain-community>=0.3.22
langchain-groq>=0.2.3
langchain-huggingface>=0.1.2
langchain-chroma>=0.2.2
chromadb>=0.6.3
pypdf>=5.4.0
python-docx>=1.1.2
docx2txt>=0.9
pydantic>=2.11.3
pydantic-settings>=2.8.1
python-dotenv>=1.1.0
tenacity>=9.0.0
numpy>=2.2.5
sentence-transformers>=3.4.1
